requirements.txt
anyio==3.5.0
asgiref==3.5.0
asyncpg==0.25.0
bcrypt==3.2.0
black==22.1.0
cffi==1.15.0
click==8.0.3
colorama==0.4.4
databases==0.5.5
dnspython==2.2.0
email-validator==1.1.3
fastapi==0.73.0
fastapi-users==9.2.3
fastapi-users-db-sqlalchemy==2.0.4
greenlet==1.1.2
h11==0.13.0
idna==3.3
Jinja2==3.0.3
loguru==0.6.0
makefun==1.13.1
MarkupSafe==2.0.1
mypy-extensions==0.4.3
passlib==1.7.4
pathspec==0.9.0
platformdirs==2.4.1
psycopg2==2.9.3
pycparser==2.21
pydantic==1.9.0
PyJWT==2.3.0
python-dotenv==0.19.2
python-multipart==0.0.5
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.31
starlette==0.17.1
tomli==2.0.0
typing_extensions==4.0.1
uvicorn==0.17.1
win32-setctime==1.1.0
